Bryan: Moved fun page to about me; fixed title
Fun page gone (migrated to about me)
Title in projects looks better
Commented out settings options that weren’t implemented

TEST=manual
