+++ js/src/personal/bryan/settings/Settings.tsx
@@ -32,8 +32,8 @@ export function Settings() {
           <Group justify="center" px="md">
             <Text>{'Change light/dark mode'}</Text>
             <LightDark />
-            <Text>{'Change font size'}</Text>
-            <Text>{'Change background'}</Text>
+            {/*<Text>{'Change font size'}</Text>*/}
+            {/*<Text>{'Change background'}</Text>*/}
           </Group>
         </HoverCard.Dropdown>
       </HoverCard>
